Moved logic for filtering organisation in reports to a mixin
